requirements.txt
pyzotero
python-dateutil
difflib